src/backend/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
pyotp==2.9.0
python-dotenv==1.0.1
Werkzeug==3.0.3
qrcode==7.4.2
Pillow==10.4.0
